Add InvalidStepDefinitionError
Resolves #78.

InvalidStepDefinitionError is raised if step definition does not respond
to `#call`. It covers both missing definitions and improper definitions
(i.e. strings).

Also, OperationResolver used `&&` to chain operations, which lead to
presence of `false` in the result. Got rid of it. Now it's either
object from the container, or a nil.