Change adapter API
This commit changes the existign adapter API, however, the changes are
essentially cosmetic. Step adapters don't get steps as input parameters anymore,
this was clearly a mixing of concerns, as in, a step calls its adapter and then
the adapter calls the step back. I added two simple wrapper classes called
StepAdater and Callable.

StepAdater simply wraps an adater object. Since we don't use inheritance (and
this is awesome) we don't have common behavior between adapters besides the
`.call` method. A StepAdapter instance stores step options and the actual
operation to be called. It also dynamically determines whether the adaper yields
a block thus make it possible to distinguish around-ish steps from before-ish
ones. To do so it relies on the list of arguments of the adapter, if it has a
block object then the adapter can control the execution. I know this might look
unnecessary smart but I think there will be not that many around-like adapters,
in fact, I think there will be exactly one so it's better to have the adapter
interface thiner.

Callable is a simple wrapper around the actual operation, it exists only for
checking operation's arity.

These classes simplified the Step class quite significantly and allowed me to
separate flows for around-like adapters without any hacks.