Restore the broadcast of success value on step_succeeded event
Because the step input args are variadic, this value is broadcast as the second argument, immediately after the step name, and before the rest of the step’s input args.

This also keeps it consistent with the arguments broadcast with the step_failed event.

Along with this, tidy up the related specs a little.