[WIP] Allow steps to take inputs from any output
This follows from [a conversation in the
forums](https://discourse.dry-rb.org/t/dry-transaction-operation-ignoring-input/670/4).

Basically, it transforms dry-transaction into a DSL for the result
monad, allowing any step to take as inputs the outputs from any other
previous step. It is controlled with the `input:` option in the step
adapter. For example:

```ruby
step :log, with: :log, input: [:process, :persist]
```

would make the `log` operation to take as positional arguments the
outputs of `:process` and `:persist` steps, along with any other
additional step argument.

Inputs order is not coupled to steps order, so you can do:

```ruby
step :log, with: :log, input: [:persist, :process]
```

You can always reference the initial transaction input with the
`:_initial` reserved word:

```ruby
step :log, with: :log, input: [:_initial]
```

An empty list corresponds to the "continue" or "then" (`>>`) monadic
operation.

If the `input:` option is not given, the behaviour defaults to be like
always: to take the last step output as input.

An alternative API could be more explicit about this being a DSL for
monads:

```ruby
step :log, with: :log, type: :bind, input: [:persist, :process]
step :log_two, with: :log_two, type: :then
```

In this case, `:type` would default to `:bind`, while `:then` would
ignore an `:input` option when it is given. This would be more explicit,
but it would add more complexity to the code and mainly to an end user
not familiar with monads.

**Important:** This commit breaks the `around:` step behaviour. If
dry-transaction becomes a DSL for monads, then there is no place for
`around:` as it is. I think we should redesign it making it something
between the transaction and the steps, but not a step adapter itself.
This way we can still have the steps being pure monadic code, while the
transaction's `around:` behaviour would deal with anything else.