Re-allow Hash-like parameters to be passed to steps
Previous commit 49c8dc24cf8e826c5771ddce5ec631f7952354a1 fixed a ruby 2.7 / ruby 3.0 issue, where you are no longer allowed to pass hash structures as the last argument to a method.

However, this fix affected other types of objects, that may inherit from Hash, and behave in a very similar way to Hash, specifically HashWithIndifferentAccess.

HashWithIndifferentAccess is used as the default object type for Rails parameters, and as such, is often passed to validation steps, as a HashWithIndifferentAccess type. The underlying validation code is often written with that implicit expectation, and may use strings or symbols as keys. This breaks, when the parameter is transformed in to a simple Hash.

While, it would probably be ideal to be able to go back, and remove HashWithIndifferentAccess from validators, and tighten the expectations to explicitly use strings or symbols, this fix was a breaking change to validators build with Rails parameters in mind.

This commit adds a test with a simple HashWithIndifferentAccess like test class, and demonstrates typical validation code. Without the included fix, these tests fail on current master (49c8dc24cf8e826c5771ddce5ec631f7952354a1) but pass on the commit before the fixes (69e886005ceff48a8f917011ceaf1bf2b90d03e9). The included fix limits the scope of the previous fix to only apply explicitly to Hash instances, leaving "hash-like" objects untouched.